requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anthropic==0.31.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
chardet==5.1.0
charset-normalizer==3.2.0
dataclasses-json==0.5.13
distro==1.8.0
exceptiongroup==1.1.2
feedparser==6.0.10
filelock==3.15.4
frozenlist==1.4.0
fsspec==2024.6.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
Jinja2==3.1.4
jiter==0.5.0
langchain==0.0.247
langsmith==0.0.15
llvmlite==0.43.0
lxml==4.9.3
MarkupSafe==2.1.5
marshmallow==3.20.1
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.60.0
numexpr==2.8.4
numpy==1.25.1
openai-whisper==20231117
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.12
pysrt==1.1.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
sympy==1.13.0
tenacity==8.2.2
tiktoken==0.7.0
tokenizers==0.13.3
torch==2.3.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.0.4
yarl==1.9.2
